Distributed animations by files
